+++ DelvUI/changelog.md
@@ -10,6 +10,7 @@ Features:
 - Added support for Trust and Squadron Command parties.
 - Added Enable Combat Hotbars sub-option to use with Cross Hotbar instead.
 - Added option to change background color and opacity of the Castbars.
+- Added an option to display mana bar for jobs with raise (in addition to healers) in party frames.
 
 Fixes:
 - Fixed unit and party frames not being interactable when using XivAlexander.